# lib/api-zod/src/generated/types/createPersonRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * People Review App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePersonRequest {
  /**
   * @minLength 2
   * @maxLength 100
   */
  name: string;
  /** @maxLength 500 */
  description?: string;
}