# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * People Review App API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateReviewRequestFlagType } from "./createReviewRequestFlagType";

export interface CreateReviewRequest {
  flagType: CreateReviewRequestFlagType;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  /**
   * @minLength 10
   * @maxLength 1000
   */
  text: string;
}